Move to e2etest
