+++ packages/@react-native-windows/cli/src/e2etest/dependencyConfig.test.ts
@@ -8,9 +8,9 @@ import * as fs from 'fs';
 import * as path from 'path';
 
 import {
-  projectConfigWindows,
-  WindowsProjectConfig,
-} from '../config/projectConfig';
+  dependencyConfigWindows,
+  WindowsDependencyConfig,
+} from '../config/dependencyConfig';
 
 import {copyAndReplace} from '../generator-common';
 
@@ -35,68 +35,68 @@ function project(
 
 const projects: TargetProject[] = [
   // Nothing but a react-native.config.js, with nulls
-  project('BlankApp'),
+  project('BlankLib'),
   // Nothing but a windows folder
-  project('MissingProjectFilesApp', async (folder: string) => {
+  project('MissingProjectFilesLib', async (folder: string) => {
     const windowsDir = path.join(folder, 'windows');
     await tryMkdir(windowsDir);
   }),
   // New C++ project based on the template
-  project('SimpleCppApp', async (folder: string) => {
+  project('SimpleCppLib', async (folder: string) => {
     const windowsDir = path.join(folder, 'windows');
     await tryMkdir(windowsDir);
 
     const replacements = {
-      name: 'SimpleCppApp',
-      namespace: 'SimpleCppApp',
+      name: 'SimpleCppLib',
+      namespace: 'SimpleCppLib',
       useMustache: true,
       projectGuidUpper: testProjectGuid,
       projectGuidLower: testProjectGuid.toLowerCase(),
     };
 
     await copyAndReplace(
-      path.join(templateRoot, 'cpp-app/proj/MyApp.sln'),
-      path.join(windowsDir, 'SimpleCppApp.sln'),
+      path.join(templateRoot, 'cpp-lib/proj/MyLib.sln'),
+      path.join(windowsDir, 'SimpleCppLib.sln'),
       replacements,
       null,
     );
 
-    const projDir = path.join(windowsDir, 'SimpleCppApp');
+    const projDir = path.join(windowsDir, 'SimpleCppLib');
     await tryMkdir(projDir);
 
     await copyAndReplace(
-      path.join(templateRoot, 'cpp-app/proj/MyApp.vcxproj'),
-      path.join(projDir, 'SimpleCppApp.vcxproj'),
+      path.join(templateRoot, 'cpp-lib/proj/MyLib.vcxproj'),
+      path.join(projDir, 'SimpleCppLib.vcxproj'),
       replacements,
       null,
     );
   }),
   // New C# project based on the template
-  project('SimpleCSharpApp', async (folder: string) => {
+  project('SimpleCSharpLib', async (folder: string) => {
     const windowsDir = path.join(folder, 'windows');
     await tryMkdir(windowsDir);
 
     const replacements = {
-      name: 'SimpleCSharpApp',
-      namespace: 'SimpleCSharpApp',
+      name: 'SimpleCSharpLib',
+      namespace: 'SimpleCSharpLib',
       useMustache: true,
       projectGuidUpper: testProjectGuid,
       projectGuidLower: testProjectGuid.toLowerCase(),
     };
 
     await copyAndReplace(
-      path.join(templateRoot, 'cs-app/proj/MyApp.sln'),
-      path.join(windowsDir, 'SimpleCSharpApp.sln'),
+      path.join(templateRoot, 'cs-lib/proj/MyLib.sln'),
+      path.join(windowsDir, 'SimpleCSharpLib.sln'),
       replacements,
       null,
     );
 
-    const projDir = path.join(windowsDir, 'SimpleCSharpApp');
+    const projDir = path.join(windowsDir, 'SimpleCSharpLib');
     await tryMkdir(projDir);
 
     await copyAndReplace(
-      path.join(templateRoot, 'cs-app/proj/MyApp.csproj'),
-      path.join(projDir, 'SimpleCSharpApp.csproj'),
+      path.join(templateRoot, 'cs-lib/proj/MyLib.csproj'),
+      path.join(projDir, 'SimpleCSharpLib.csproj'),
       replacements,
       null,
     );
@@ -105,62 +105,63 @@ const projects: TargetProject[] = [
 
 // Tests that given userConfig is null, the result will always be null
 test.each(projects)(
-  'projectConfig - %s (userConfig is null)',
+  'dependencyConfig - %s (userConfig is null)',
   async (name, setup) => {
-    const folder = path.resolve('src/test/projects/', name);
+    const folder = path.resolve('src/e2etest/projects/', name);
 
     if (setup !== undefined) {
       await setup(folder);
     }
 
     const userConfig = null;
-    const expectedConfig: WindowsProjectConfig | null = null;
+    const expectedConfig: WindowsDependencyConfig | null = null;
 
-    expect(projectConfigWindows(folder, userConfig)).toBe(expectedConfig);
+    expect(dependencyConfigWindows(folder, userConfig)).toBe(expectedConfig);
   },
 );
 
 // Tests the result given a windows project config in react-native.config.js
 test.each(projects)(
-  'projectConfig - %s (Use react-native.config.js)',
+  'dependencyConfig - %s (Use react-native.config.js)',
   async (name, setup) => {
-    const folder = path.resolve('src/test/projects/', name);
+    const folder = path.resolve('src/e2etest/projects/', name);
     const rnc = require(path.join(folder, 'react-native.config.js'));
 
     if (setup !== undefined) {
       await setup(folder);
     }
 
-    const userConfig: Partial<WindowsProjectConfig> = rnc.project.windows;
-    let expectedConfig: WindowsProjectConfig | null = rnc.expectedConfig;
+    const userConfig: Partial<WindowsDependencyConfig> =
+      rnc.dependency.platforms.windows;
+    let expectedConfig: WindowsDependencyConfig | null = rnc.expectedConfig;
 
     if (expectedConfig !== null) {
       expectedConfig.folder = folder;
     }
 
-    expect(projectConfigWindows(folder, userConfig)).toEqual(expectedConfig);
+    expect(dependencyConfigWindows(folder, userConfig)).toEqual(expectedConfig);
   },
 );
 
 // Tests the result of ignoring the windows project config in react-native.config.js
 test.each(projects)(
-  'projectConfig - %s (Ignore react-native.config.js)',
+  'dependencyConfig - %s (Ignore react-native.config.js)',
   async (name, setup) => {
-    const folder = path.resolve('src/test/projects/', name);
+    const folder = path.resolve('src/e2etest/projects/', name);
     const rnc = require(path.join(folder, 'react-native.config.js'));
 
     if (setup !== undefined) {
       await setup(folder);
     }
 
-    const userConfig: Partial<WindowsProjectConfig> = {};
-    const expectedConfig: WindowsProjectConfig | null =
+    const userConfig: Partial<WindowsDependencyConfig> = {};
+    const expectedConfig: WindowsDependencyConfig | null =
       rnc.expectedConfigIgnoringOverride;
 
     if (expectedConfig !== null) {
       expectedConfig.folder = folder;
     }
 
-    expect(projectConfigWindows(folder, userConfig)).toEqual(expectedConfig);
+    expect(dependencyConfigWindows(folder, userConfig)).toEqual(expectedConfig);
   },
 );
