+++ packages/@react-native-windows/cli/src/e2etest/projectConfig.test.ts
@@ -8,9 +8,9 @@ import * as fs from 'fs';
 import * as path from 'path';
 
 import {
-  dependencyConfigWindows,
-  WindowsDependencyConfig,
-} from '../config/dependencyConfig';
+  projectConfigWindows,
+  WindowsProjectConfig,
+} from '../config/projectConfig';
 
 import {copyAndReplace} from '../generator-common';
 
@@ -35,68 +35,68 @@ function project(
 
 const projects: TargetProject[] = [
   // Nothing but a react-native.config.js, with nulls
-  project('BlankLib'),
+  project('BlankApp'),
   // Nothing but a windows folder
-  project('MissingProjectFilesLib', async (folder: string) => {
+  project('MissingProjectFilesApp', async (folder: string) => {
     const windowsDir = path.join(folder, 'windows');
     await tryMkdir(windowsDir);
   }),
   // New C++ project based on the template
-  project('SimpleCppLib', async (folder: string) => {
+  project('SimpleCppApp', async (folder: string) => {
     const windowsDir = path.join(folder, 'windows');
     await tryMkdir(windowsDir);
 
     const replacements = {
-      name: 'SimpleCppLib',
-      namespace: 'SimpleCppLib',
+      name: 'SimpleCppApp',
+      namespace: 'SimpleCppApp',
       useMustache: true,
       projectGuidUpper: testProjectGuid,
       projectGuidLower: testProjectGuid.toLowerCase(),
     };
 
     await copyAndReplace(
-      path.join(templateRoot, 'cpp-lib/proj/MyLib.sln'),
-      path.join(windowsDir, 'SimpleCppLib.sln'),
+      path.join(templateRoot, 'cpp-app/proj/MyApp.sln'),
+      path.join(windowsDir, 'SimpleCppApp.sln'),
       replacements,
       null,
     );
 
-    const projDir = path.join(windowsDir, 'SimpleCppLib');
+    const projDir = path.join(windowsDir, 'SimpleCppApp');
     await tryMkdir(projDir);
 
     await copyAndReplace(
-      path.join(templateRoot, 'cpp-lib/proj/MyLib.vcxproj'),
-      path.join(projDir, 'SimpleCppLib.vcxproj'),
+      path.join(templateRoot, 'cpp-app/proj/MyApp.vcxproj'),
+      path.join(projDir, 'SimpleCppApp.vcxproj'),
       replacements,
       null,
     );
   }),
   // New C# project based on the template
-  project('SimpleCSharpLib', async (folder: string) => {
+  project('SimpleCSharpApp', async (folder: string) => {
     const windowsDir = path.join(folder, 'windows');
     await tryMkdir(windowsDir);
 
     const replacements = {
-      name: 'SimpleCSharpLib',
-      namespace: 'SimpleCSharpLib',
+      name: 'SimpleCSharpApp',
+      namespace: 'SimpleCSharpApp',
       useMustache: true,
       projectGuidUpper: testProjectGuid,
       projectGuidLower: testProjectGuid.toLowerCase(),
     };
 
     await copyAndReplace(
-      path.join(templateRoot, 'cs-lib/proj/MyLib.sln'),
-      path.join(windowsDir, 'SimpleCSharpLib.sln'),
+      path.join(templateRoot, 'cs-app/proj/MyApp.sln'),
+      path.join(windowsDir, 'SimpleCSharpApp.sln'),
       replacements,
       null,
     );
 
-    const projDir = path.join(windowsDir, 'SimpleCSharpLib');
+    const projDir = path.join(windowsDir, 'SimpleCSharpApp');
     await tryMkdir(projDir);
 
     await copyAndReplace(
-      path.join(templateRoot, 'cs-lib/proj/MyLib.csproj'),
-      path.join(projDir, 'SimpleCSharpLib.csproj'),
+      path.join(templateRoot, 'cs-app/proj/MyApp.csproj'),
+      path.join(projDir, 'SimpleCSharpApp.csproj'),
       replacements,
       null,
     );
@@ -105,63 +105,62 @@ const projects: TargetProject[] = [
 
 // Tests that given userConfig is null, the result will always be null
 test.each(projects)(
-  'dependencyConfig - %s (userConfig is null)',
+  'projectConfig - %s (userConfig is null)',
   async (name, setup) => {
-    const folder = path.resolve('src/test/projects/', name);
+    const folder = path.resolve('src/e2etest/projects/', name);
 
     if (setup !== undefined) {
       await setup(folder);
     }
 
     const userConfig = null;
-    const expectedConfig: WindowsDependencyConfig | null = null;
+    const expectedConfig: WindowsProjectConfig | null = null;
 
-    expect(dependencyConfigWindows(folder, userConfig)).toBe(expectedConfig);
+    expect(projectConfigWindows(folder, userConfig)).toBe(expectedConfig);
   },
 );
 
 // Tests the result given a windows project config in react-native.config.js
 test.each(projects)(
-  'dependencyConfig - %s (Use react-native.config.js)',
+  'projectConfig - %s (Use react-native.config.js)',
   async (name, setup) => {
-    const folder = path.resolve('src/test/projects/', name);
+    const folder = path.resolve('src/e2etest/projects/', name);
     const rnc = require(path.join(folder, 'react-native.config.js'));
 
     if (setup !== undefined) {
       await setup(folder);
     }
 
-    const userConfig: Partial<WindowsDependencyConfig> =
-      rnc.dependency.platforms.windows;
-    let expectedConfig: WindowsDependencyConfig | null = rnc.expectedConfig;
+    const userConfig: Partial<WindowsProjectConfig> = rnc.project.windows;
+    let expectedConfig: WindowsProjectConfig | null = rnc.expectedConfig;
 
     if (expectedConfig !== null) {
       expectedConfig.folder = folder;
     }
 
-    expect(dependencyConfigWindows(folder, userConfig)).toEqual(expectedConfig);
+    expect(projectConfigWindows(folder, userConfig)).toEqual(expectedConfig);
   },
 );
 
 // Tests the result of ignoring the windows project config in react-native.config.js
 test.each(projects)(
-  'dependencyConfig - %s (Ignore react-native.config.js)',
+  'projectConfig - %s (Ignore react-native.config.js)',
   async (name, setup) => {
-    const folder = path.resolve('src/test/projects/', name);
+    const folder = path.resolve('src/e2etest/projects/', name);
     const rnc = require(path.join(folder, 'react-native.config.js'));
 
     if (setup !== undefined) {
       await setup(folder);
     }
 
-    const userConfig: Partial<WindowsDependencyConfig> = {};
-    const expectedConfig: WindowsDependencyConfig | null =
+    const userConfig: Partial<WindowsProjectConfig> = {};
+    const expectedConfig: WindowsProjectConfig | null =
       rnc.expectedConfigIgnoringOverride;
 
     if (expectedConfig !== null) {
       expectedConfig.folder = folder;
     }
 
-    expect(dependencyConfigWindows(folder, userConfig)).toEqual(expectedConfig);
+    expect(projectConfigWindows(folder, userConfig)).toEqual(expectedConfig);
   },
 );
