+++ packages/@react-native-windows/cli/src/e2etest/projects/BlankApp/react-native.config.js
