+++ packages/@react-native-windows/cli/src/e2etest/projects/BlankLib/react-native.config.js
