+++ packages/@react-native-windows/cli/src/e2etest/projects/MissingProjectFilesApp/.gitignore
