+++ packages/@react-native-windows/cli/src/e2etest/projects/MissingProjectFilesLib/.gitignore
