+++ packages/@react-native-windows/cli/src/e2etest/projects/MissingProjectFilesLib/react-native.config.js
