+++ packages/@react-native-windows/cli/src/e2etest/projects/SimpleCSharpApp/react-native.config.js
