+++ packages/@react-native-windows/cli/src/e2etest/projects/SimpleCSharpLib/react-native.config.js
