+++ packages/@react-native-windows/cli/src/e2etest/projects/SimpleCppApp/react-native.config.js
