+++ packages/@react-native-windows/cli/src/e2etest/projects/SimpleCppLib/.gitignore
