+++ packages/@react-native-windows/cli/src/e2etest/projects/SimpleCppLib/react-native.config.js
